# [perf] Speed up align sum kernels