--- a/csrc/moe/moe_align_sum_kernels.cu
+++ b/csrc/moe/moe_align_sum_kernels.cu
@@ -223,7 +223,7 @@ void moe_align_block_size(torch::Tensor topk_ids, int64_t num_experts,
 
   // BlockScan uses 1024 threads and assigns one thread per expert.
   TORCH_CHECK(padded_num_experts < 1024,
-              "padded_num_experts must be less than 1024"); 
+              "padded_num_experts must be less than 1024");
 
   VLLM_DISPATCH_INTEGRAL_AND_UNSIGNED_TYPES(
       topk_ids.scalar_type(), "moe_align_block_size_kernel", [&] {