## Purpose
- Move fill ops inside align sum kernel to reduce bubbles.
- cumsum buffer does not need to be filled with zero.
- we can use blockscan to do the prefix sum

This PR also moves the triton inits into the kernel to make it a fair comparison and also ensure the kernel is usable in the future as a fallback if required.

## Benchmarks
### Main branch
```
FP16:
# vllm bench throughput --model Qwen/Qwen3-30B-A3B --load-format dummy --input-len 1000 --output-len 100

Throughput: 43.75 requests/s, 48024.34 total tokens/s, 4374.91 output tokens/s
Total num prompt tokens:  997723
Total num output tokens:  100000

FP8:
# vllm bench throughput --model Qwen/Qwen3-30B-A3B-FP8 --load-format dummy --input-len 1000 --output-len 100

Throughput: 41.04 requests/s, 45049.17 total tokens/s, 4103.87 output tokens/s
Total num prompt tokens:  997723
Total num output tokens:  100000

Kernel benchmark:
# python3 benchmarks/kernels/benchmark_moe_align_block_size.py

Running correctness check...
✅ VLLM implementation works with 64 experts!
✅ Triton and VLLM implementations match.
moe-align-block-size-performance:
    num_tokens  num_experts  topk       VLLM      Triton
0          1.0         16.0   1.0  16.448000   23.040000
1          1.0         16.0   2.0  16.432000   23.104001
2          1.0         16.0   8.0  16.448000   23.040000
3          1.0         64.0   1.0  21.600001   25.984000
4          1.0         64.0   2.0  21.792000   26.048001
5          1.0         64.0   8.0  21.824000   25.952000
6          1.0        224.0   1.0  23.680000   40.288001
7          1.0        224.0   2.0  23.680000   40.320002
8          1.0        224.0   8.0  23.712000   40.383998
9          1.0        256.0   1.0  24.607999   43.136001
10         1.0        256.0   2.0  24.639999   43.104000
11         1.0        256.0   8.0  24.639999   43.200001
12         1.0        280.0   1.0  25.248000   45.407999
13         1.0        280.0   2.0  25.248000   45.343999
14         1.0        280.0   8.0  25.248000   45.440000
15         1.0        512.0   1.0  31.136001   69.151998
16         1.0        512.0   2.0  31.328000   69.119997
17         1.0        512.0   8.0  31.296000   69.215998
18        16.0         16.0   1.0  16.511999   23.296000
19        16.0         16.0   2.0  16.608000   23.520000
20        16.0         16.0   8.0  17.856000   24.351999
21        16.0         64.0   1.0  21.792000   26.400000
22        16.0         64.0   2.0  21.792000   26.656000
23        16.0         64.0   8.0  22.143999   27.424000
24        16.0        224.0   1.0  23.871999   41.503999
25        16.0        224.0   2.0  23.903999   41.600000
26        16.0        224.0   8.0  24.032000   41.152000
27        16.0        256.0   1.0  24.768000   43.088000
28        16.0        256.0   2.0  24.831999   43.136001
29        16.0        256.0   8.0  24.928000   43.391999
30        16.0        280.0   1.0  25.152000   45.968000
31        16.0        280.0   2.0  25.184000   46.080001
32        16.0        280.0   8.0  25.343999   46.271998
33        16.0        512.0   1.0  31.264000   69.343999
34        16.0        512.0   2.0  31.328000   69.504000
35        16.0        512.0   8.0  31.456001   69.888003
36       256.0         16.0   1.0  19.200001   25.312001
37       256.0         16.0   2.0  22.624001   28.576000
38       256.0         16.0   8.0  18.528000   45.184001
39       256.0         64.0   1.0  23.104001   28.416000
40       256.0         64.0   2.0  24.831999   29.023999
41       256.0         64.0   8.0  20.256000   33.535998
42       256.0        224.0   1.0  24.256000   42.367999
43       256.0        224.0   2.0  24.000000   42.943999
44       256.0        224.0   8.0  24.256000   45.952000
45       256.0        256.0   1.0  25.119999   44.224001
46       256.0        256.0   2.0  24.960000   44.192001
47       256.0        256.0   8.0  25.984000   47.488000
48       256.0        280.0   1.0  25.312001   46.239998
49       256.0        280.0   2.0  25.536001   47.327999
50       256.0        280.0   8.0  26.432000   49.568001
51       256.0        512.0   1.0  31.488001   69.824003
52       256.0        512.0   2.0  31.392001   69.856003
53       256.0        512.0   8.0  32.671999   71.712002
54      4096.0         16.0   1.0  20.128001   68.896003
55      4096.0         16.0   2.0  22.720000  114.367999
56      4096.0         16.0   8.0  36.256000  378.015995
57      4096.0         64.0   1.0  21.856001   39.391998
58      4096.0         64.0   2.0  24.639999   51.872000
59      4096.0         64.0   8.0  41.216001  121.360000
60      4096.0        224.0   1.0  26.368000   50.976001
61      4096.0        224.0   2.0  29.023999   56.607999
62      4096.0        224.0   8.0  45.504000   78.304000
63      4096.0        256.0   1.0  27.071999   51.968001
64      4096.0        256.0   2.0  29.824000   58.944002
65      4096.0        256.0   8.0  45.568001   78.368001
66      4096.0        280.0   1.0  27.295999   53.056002
67      4096.0        280.0   2.0  30.272000   59.648000
68      4096.0        280.0   8.0  43.264002   80.095999
69      4096.0        512.0   1.0  33.824001   73.600002
70      4096.0        512.0   2.0  35.551999   77.776000
71      4096.0        512.0   8.0  49.024001   98.591998
```

### This PR
```
FP16:
#vllm bench throughput --model Qwen/Qwen3-30B-A3B --load-format dummy --input-len 1000 --output-len 100

Throughput: 43.94 requests/s, 48234.94 total tokens/s, 4394.09 output tokens/s
Total num prompt tokens:  997723
Total num output tokens:  100000

FP8:
#vllm bench throughput --model Qwen/Qwen3-30B-A3B-FP8 --load-format dummy --input-len 1000 --output-len 100

Throughput: 41.26 requests/s, 45294.95 total tokens/s, 4126.26 output tokens/s
Total num prompt tokens:  997723
Total num output tokens:  100000

Kernel benchmark:
# python3 benchmarks/kernels/benchmark_moe_align_block_size.py

Running correctness check...
✅ VLLM implementation works with 64 experts!
✅ Triton and VLLM implementations match.
moe-align-block-size-performance:
    num_tokens  num_experts  topk       VLLM      Triton
0          1.0         16.0   1.0  17.472001   27.488001
1          1.0         16.0   2.0  17.600000   30.304000
2          1.0         16.0   8.0  17.696001   30.880000
3          1.0         64.0   1.0  25.760001   31.296000
4          1.0         64.0   2.0  25.855999   31.168001
5          1.0         64.0   8.0  25.823999   31.488001
6          1.0        224.0   1.0  21.536000   44.544000
7          1.0        224.0   2.0  21.344000   44.799998
8          1.0        224.0   8.0  21.407999   44.736002
9          1.0        256.0   1.0  22.080000   47.616001
10         1.0        256.0   2.0  21.568000   47.392000
11         1.0        256.0   8.0  21.760000   47.711998
12         1.0        280.0   1.0  21.952000   49.632002
13         1.0        280.0   2.0  22.336001   49.984001
14         1.0        280.0   8.0  22.048000   49.952000
15         1.0        512.0   1.0  25.888000   75.071998
16         1.0        512.0   2.0  25.952000   75.328000
17         1.0        512.0   8.0  25.952000   75.007997
18        16.0         16.0   1.0  17.600000   27.295999
19        16.0         16.0   2.0  17.600000   28.352000
20        16.0         16.0   8.0  18.912001   29.696001
21        16.0         64.0   1.0  25.696000   31.184000
22        16.0         64.0   2.0  25.632000   30.688001
23        16.0         64.0   8.0  25.952000   30.944001
24        16.0        224.0   1.0  21.312000   45.855999
25        16.0        224.0   2.0  21.183999   45.791999
26        16.0        224.0   8.0  21.536000   45.440000
27        16.0        256.0   1.0  21.792000   47.359999
28        16.0        256.0   2.0  21.760000   47.584001
29        16.0        256.0   8.0  21.760000   47.807999
30        16.0        280.0   1.0  22.048000   50.271999
31        16.0        280.0   2.0  21.888001   50.464001
32        16.0        280.0   8.0  22.336001   50.624002
33        16.0        512.0   1.0  25.664000   74.975997
34        16.0        512.0   2.0  25.696000   75.039998
35        16.0        512.0   8.0  25.952000   75.135998
36       256.0         16.0   1.0  20.320000   29.088000
37       256.0         16.0   2.0  23.871999   32.543998
38       256.0         16.0   8.0  17.600000   49.279999
39       256.0         64.0   1.0  26.784001   32.448001
40       256.0         64.0   2.0  28.384000   32.127999
41       256.0         64.0   8.0  18.912001   37.535999
42       256.0        224.0   1.0  21.536000   46.720002
43       256.0        224.0   2.0  21.695999   47.488000
44       256.0        224.0   8.0  21.856001   50.175998
45       256.0        256.0   1.0  22.336001   48.703998
46       256.0        256.0   2.0  21.952000   48.351999
47       256.0        256.0   8.0  23.072001   51.711999
48       256.0        280.0   1.0  22.240000   50.783999
49       256.0        280.0   2.0  22.752000   52.000001
50       256.0        280.0   8.0  23.808001   54.639999
51       256.0        512.0   1.0  26.208000   75.744003
52       256.0        512.0   2.0  26.335999   75.103998
53       256.0        512.0   8.0  26.656000   77.215999
54      4096.0         16.0   1.0  19.168001   72.672002
55      4096.0         16.0   2.0  22.112001  117.183998
56      4096.0         16.0   8.0  37.087999  382.703990
57      4096.0         64.0   1.0  20.352000   43.423999
58      4096.0         64.0   2.0  23.424000   55.712000
59      4096.0         64.0   8.0  42.016000  125.568002
60      4096.0        224.0   1.0  23.264000   55.744000
61      4096.0        224.0   2.0  26.912000   60.864002
62      4096.0        224.0   8.0  44.704001   81.919998
63      4096.0        256.0   1.0  24.383999   56.448001
64      4096.0        256.0   2.0  27.327999   63.104004
65      4096.0        256.0   8.0  44.319998   82.496002
66      4096.0        280.0   1.0  23.808001   57.824001
67      4096.0        280.0   2.0  27.424000   64.576000
68      4096.0        280.0   8.0  41.792002   83.967999
69      4096.0        512.0   1.0  27.744001   79.135999
70      4096.0        512.0   2.0  30.479999   83.328001
71      4096.0        512.0   8.0  45.536000  103.808001

```

## Test Result
- `pytest tests/kernels/moe/test_moe_align_block_size.py` - PASSED

## (Optional) Documentation Update

<!--- pyml disable-next-line no-emphasis-as-heading -->