@hj-mistral 2025-07-17
> Thanks for the work! Could you also please benchmark the performance (E2E throughput + kernel latency) and make sure all unit test passes?

Any documentation to follow on how to run both?