@hj-mistral 2025-07-18
> > Any documentation to follow on how to run both?
> 
> ```shell
> Throughput(fp16)
> 
> vllm bench throughput --model Qwen/Qwen3-30B-A3B --load-format dummy --input-len 1000 --output-len 100
> 
> Throughput(fp8)
> 
> vllm bench throughput --model Qwen/Qwen3-30B-A3B-FP8 --load-format dummy --input-len 1000 --output-len 100
> ```
> 
> `vllm-source/benchmarks/kernels/benchmark_moe_align_block_size.py`
> 
> `vllm-source/tests/kernels/moe/test_moe_align_block_size.py`

All done and added to description, ptal :)