@mgoin 2025-07-19
That's a great point and answers my question as well. Don't worry about the DCO as we can resolve it manually before merge. It looks like there are a few related failures in the kernel tests