@tdoublep 2025-07-22
The changes from this PR are causing illegal memory accesses for me. 

If I deploy with commit before this PR was merged `005ae9be6c22dfa2c2c5580b50b41e67faee4a87`:
```
$ VLLM_USE_V1=1 VLLM_ATTENTION_BACKEND=FLASHINFER vllm serve ibm-granite/granite-4.0-tiny-preview --no-enable-prefix-caching
...
INFO:     Started server process [604208]
INFO:     Waiting for application startup.
INFO:     Application startup complete.
```
Whereas, if I deploy at commit after this PR was merged `0ec82edda59aaf5cf3b07aadf4ecce1aa1131add`:
```
$ VLLM_USE_V1=1 VLLM_ATTENTION_BACKEND=FLASHINFER vllm serve ibm-granite/granite-4.0-tiny-preview --no-enable-prefix-caching
...
  File "/home/zrltpa/vllm/vllm/model_executor/layers/fused_moe/fused_moe.py", line 1230, in torch_vllm_inplace_fused_experts
    torch.ops.vllm.inplace_fused_experts(**kwargs)
  File "/home/zrltpa/miniforge3/envs/dev-env/lib/python3.12/site-packages/torch/_ops.py", line 1158, in __call__
    return self._op(*args, **(kwargs or {}))
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/zrltpa/vllm/vllm/model_executor/layers/fused_moe/fused_moe.py", line 1020, in inplace_fused_experts
    fused_experts_impl(hidden_states, w1, w2, topk_weights, topk_ids, True,
  File "/home/zrltpa/vllm/vllm/model_executor/layers/fused_moe/fused_moe.py", line 1484, in fused_experts_impl
    invoke_fused_moe_kernel(qcurr_hidden_states,
  File "/home/zrltpa/vllm/vllm/model_executor/layers/fused_moe/fused_moe.py", line 604, in invoke_fused_moe_kernel
    fused_moe_kernel[grid](
  File "/home/zrltpa/miniforge3/envs/dev-env/lib/python3.12/site-packages/triton/runtime/jit.py", line 347, in <lambda>
    return lambda *args, **kwargs: self.run(grid=grid, warmup=False, *args, **kwargs)
                                   ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/zrltpa/miniforge3/envs/dev-env/lib/python3.12/site-packages/triton/runtime/jit.py", line 591, in run
    kernel.run(grid_0, grid_1, grid_2, stream, kernel.function, kernel.packed_metadata,
  File "/home/zrltpa/miniforge3/envs/dev-env/lib/python3.12/site-packages/triton/backends/nvidia/driver.py", line 529, in __call__
    self.launch(gridX, gridY, gridZ, stream, function, self.launch_cooperative_grid, global_scratch, *args)
RuntimeError: Triton Error [CUDA]: an illegal memory access was encountered
```

Could we perhaps revert the changes from this PR until we figure out what is going on here? cc @mgoin @tlrmchlsmth