@tdoublep 2025-07-22
OK the reason it passes in CI is that vLLM bumped torch version which in turn bumped Triton version to 3.3.1. That seems to resolve the error that I am seeing. Still a bit weird though? Illegal memory access in 3.3.0 but works fine in 3.3.1?